function that prints all the elements of a list_t list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,5 +1,7 @@
 #include "lists.h"
 #include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
 
 /**
  * print_list - Prints all the elements of a list_t list.
@@ -10,7 +12,7 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t nodes = 0;
+	int nodes = 0;
 
 	while (h)
 	{
